init ember-addon setup
