+++ .travis.yml
@@ -1,16 +1,24 @@
+---
 language: node_js
+node_js:
+  - "4"
+
 sudo: false
-branches: 
-  only: 
-    - master
-    - pre
-node_js: 
-  - "0.12"
-  - "0.11"
-  - "0.10"
-  - "iojs"
+
+cache:
+  directories:
+    - node_modules
+
 before_install:
-  - npm install -g grunt-cli
-  - npm install -g testem
+  - npm config set spin false
   - npm install -g bower
+  - bower --version
+  - npm install phantomjs-prebuilt
+  - node_modules/phantomjs-prebuilt/bin/phantomjs --version
+
+install:
+  - npm install
   - bower install
+
+script:
+  - npm test
