--- grunt/concat.js
@@ -1,6 +1,0 @@
-module.exports = {
-  dist: {
-    src: 'tmp/transpile/lib/**/*.amd.js',
-    dest: 'dist/<%= pkg.name %>.amd.js'
-  }
-}
