--- grunt/copy.js
@@ -1,7 +1,0 @@
-module.exports = {
-  main: {
-    files: [
-      { expand: true, flatten: true, src: ['<%= lib %>/*.js'], dest: 'dist/' }
-    ]
-  }
-}
