--- grunt/testem.js
@@ -1,10 +1,0 @@
-module.exports = {
-  env1: {
-    options: {
-      "framework": "jasmine2",
-      "test_page": "tests/test.html",
-      "launch_in_dev": ["PhantomJS", "Chrome"],
-      "launch_in_ci": ["PhantomJS"] 
-    }
-  }
-}
