+++ package.json
@@ -3,10 +3,10 @@
   "version": "1.1.0",
   "description": "OAuth2 library for Emberjs that stores tokens in the browsers localStorage",
   "homepage": "https://github.com/amkirwan/ember-ouath2",
-  "main": "dist/ember.oauth2.js",
-  "repository": {
-    "type": "git",
-    "url": "git://github.com/amkirwan/ember-oauth2.git"
+  "private": true,
+  "directories": {
+    "doc": "doc",
+    "test": "tests"
   },
   "keywords": [
     "ember-oauth2",
@@ -21,26 +21,41 @@
   "bugs": {
     "url": "https://github.com/amkirwan/ember-oauth2/issues"
   },
-  "dependencies": {},
-  "devDependencies": {
-    "grunt": "~0.4.2",
-    "load-grunt-tasks": "*",
-    "grunt-contrib-copy": "~0.8.0",
-    "grunt-contrib-jshint": "~0.11.2",
-    "grunt-contrib-uglify": "~0.9.1",
-    "grunt-contrib-jasmine": "~0.9.0",
-    "grunt-versioner": "~0.1.10",
-    "yuidocjs": "~0.8.1",
-    "grunt-es6-module-transpiler": "~0.6.0",
-    "grunt-contrib-concat": "~0.5.0",
-    "grunt-contrib-clean": "~0.6.0",
-    "load-grunt-config": "~0.17.2",
-    "testem": "~0.9.1",
-    "grunt-testem-mincer": "~0.5.18",
-    "install": "~0.1.7",
-    "grunt-preprocess": "~5.0.0"
-  },
   "scripts": {
-    "test": "grunt --verbose test-travis"
+    "build": "ember build",
+    "start": "ember server",
+    "test": "ember test"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git://github.com/amkirwan/ember-oauth2.git"
+  },
+  "engines": {
+    "node": ">= 0.10.0"
+  },
+  "author": "",
+  "license": "MIT",
+  "devDependencies": {
+    "broccoli-asset-rev": "^2.4.2",
+    "ember-ajax": "^2.0.1",
+    "ember-cli": "2.8.0",
+    "ember-cli-app-version": "^1.0.0",
+    "ember-cli-babel": "^5.1.6",
+    "ember-cli-dependency-checker": "^1.2.0",
+    "ember-cli-htmlbars": "^1.0.3",
+    "ember-cli-htmlbars-inline-precompile": "^0.3.1",
+    "ember-cli-inject-live-reload": "^1.4.0",
+    "ember-cli-jshint": "^1.0.0",
+    "ember-cli-qunit": "^2.1.0",
+    "ember-cli-release": "^0.2.9",
+    "ember-cli-sri": "^2.1.0",
+    "ember-cli-test-loader": "^1.1.0",
+    "ember-cli-uglify": "^1.2.0",
+    "ember-data": "^2.8.0",
+    "ember-export-application-global": "^1.0.5",
+    "ember-load-initializers": "^0.5.1",
+    "ember-resolver": "^2.0.3",
+    "ember-welcome-page": "^1.0.1",
+    "loader.js": "^4.0.1"
   }
 }
