--- testem.json
@@ -1,6 +1,0 @@
-{
-  "framework": "jasmine2",
-  "test_page": "tests/test.html",
-  "launch_in_dev": ["PhantomJS", "Chrome"],
-  "launch_in_ci": ["PhantomJS"]
-}
