+++ tests/.jshintrc
@@ -1,24 +1,32 @@
 {
-  "predef": {
-    "document": true,
-    "window": true,
-    "Ember": true,
-    "Em": true,
-    "$": true,
-    "App": true,
-    "expect": true,
-    "require": true,
-    "describe": true,
-    "it": true,
-    "beforeEach": true,
-    "afterEach": true,
-    "sinon": true,
-    "localStorage": true
-  },
-  "node" : false,
-  "browser" : false,
-  "boss" : true,
-  "curly": false,
+  "predef": [
+    "document",
+    "window",
+    "location",
+    "setTimeout",
+    "$",
+    "-Promise",
+    "define",
+    "console",
+    "visit",
+    "exists",
+    "fillIn",
+    "click",
+    "keyEvent",
+    "triggerEvent",
+    "find",
+    "findWithAssert",
+    "wait",
+    "DS",
+    "andThen",
+    "currentURL",
+    "currentPath",
+    "currentRouteName"
+  ],
+  "node": false,
+  "browser": false,
+  "boss": true,
+  "curly": true,
   "debug": false,
   "devel": false,
   "eqeqeq": true,
@@ -39,5 +47,6 @@
   "strict": false,
   "white": false,
   "eqnull": true,
-  "esnext": true
+  "esversion": 6,
+  "unused": true
 }
