--- tests/test.html
@@ -1,53 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <title>App</title>
-  <meta name="description" content="">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-
-    <script src="/bower_components/jquery/dist/jquery.js"></script>
-    <script src="/bower_components/handlebars/handlebars.js"></script>
-    <script src="/bower_components/ember/ember.js"></script>
-    <script src="/bower_components/handlebars/handlebars.runtime.js"></script>
-    <script src="/bower_components/sinonjs/sinon.js"></script>
-    
-    <script src="/bower_components/loader.js/loader.js"></script>
-
-    <script src="/bower_components/ember-resolver/dist/ember-resolver.js"></script>
-    <script src="/bower_components/ember-cli-shims/app-shims.js"></script>
-
-    <script src="/dist/ember-oauth2.amd.js"></script>
-
-  <link rel="stylesheet" href="/bower_components/jasmine/lib/jasmine-core/jasmine.css">
-  <script src="/bower_components/jasmine/lib/jasmine-core/jasmine.js"></script>
-  <script src="/bower_components/jasmine/lib/jasmine-core/jasmine-html.js"></script>
-  <script src="/bower_components/jasmine/lib/jasmine-core/boot.js"></script>
-  <style>
-    #ember-testing-container {
-      position: absolute;
-      background: white;
-      bottom: 0;
-      right: 0;
-      width: 640px;
-      height: 384px;
-      overflow: auto;
-      z-index: 9999;
-      border: 1px solid #ccc;
-    }
-    #ember-testing {
-      zoom: 50%;
-    }
-  </style>
-  <!-- @endif -->
-
-</head>
-<body>
-  <script src="/tests/ember-oauth2-amd.spec.js"></script>
-
-  <script src="/testem.js"></script>
-
-</body>
-</html>
-
